backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
httpx==0.27.0
litellm==1.48.0
boto3==1.35.0
pyodbc==5.1.0
pydantic==2.8.0
websockets==13.0
python-multipart==0.0.9
